requirements.txt
# Data & ML
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
matplotlib>=3.8
seaborn>=0.13
pyarrow>=16
scipy>=1.13
xgboost>=2.1
joblib>=1.4

# API
fastapi>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9
aiofiles>=23.2

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9

# Auth
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7

# Worker / Scheduling
apscheduler>=3.10
httpx>=0.27

# Notifications
python-telegram-bot>=21.0

# Utilities
pydantic>=2.0
python-dotenv>=1.0
